Create Data Processor App
Here I detail how to build your own App in Python using the wonderful technology of Streamlit. Buidling this app is great practice for if you want to become a data scientist. It guides you through the statistical process - which is a process that is hard to find online - and also teaches you how to build your very own application.